+++ client/sagas/index.js
@@ -1,12 +1,9 @@
 import { fork, all } from 'redux-saga/effects';
 
-// import auth from './authSagas';
+import configuration from './configurationSagas';
 import categories from './categorySagas';
 
-const sagas = [
-  // ...auth,
-  ...categories,
-];
+const sagas = [...categories, ...configuration];
 
 export default function* root() {
   yield all(sagas.map(saga => fork(saga)));